KRACOEUS-8032 Fix to binding editors for prop science keyword and compliance, compliance binding fix also fixes save issues described in KRACOEUS-8008
